refactor(tui): retire tests.rs into per-module test suites
`src/tui/tests.rs` was two things at once: a shared fixture module imported by
nine screen/render files, and a bag of 217 tests still grouped by screen inside
one 3,916-line file. Changing a screen meant looking for its tests in two
places.

Fixtures move to `tui::test_support`, and every test moves to the module that
owns what it asserts — behaviour tests to the screen whose state they drive,
cross-screen dispatch (quit latch, top-bar clicks, scroll, theme, yank) to
`keys.rs`, structural gates to the module they read. `src/tui/tests.rs` and the
seven `#[cfg(test)] pub(crate) use` re-exports that existed only to feed its
`use super::*` are gone.

Pure relocation: no production interface, behaviour, or assertion changed, and
the 711 lib test names are byte-identical before and after.

Closes #381
